sql: Make INT an alias for INT8
This is the first in a series of changes to allow INT to eventually be
redefined as INT4.

Currently, a "naked" INT is generally treated as a 64-bit int type, but INT is
separate from INT8. This change eliminates INT as a distinct type in favor of
always using INT8. This is analogous to how we treat FLOAT/FLOAT8.

The coltypes.Int global has been removed and we should no longer produce type
objects that have a zero Width. The existing interpretation of zero-Width types
as 64-bit values will continue indefinitely to preserve backwards-compatibility
with existing descriptors.

In general, the changes to tests are due to the widespread use of INT as an
arbitary datatype. Except in a few cases where we are testing how "INT" is
interpreted as a datatype, the tests have been changed to simply use INT8, to
avoid another major round of text fixups as work on #26925 continues.

Release note (sql change): The INT type is now treated as an alias for INT8.

+++ pkg/sql/coltypes/conv.go
@@ -73,7 +73,7 @@ func DatumTypeToColumnType(t types.T) (T, error) {
 	case types.BitArray:
 		return VarBit, nil
 	case types.Int:
-		return Int, nil
+		return Int8, nil
 	case types.Float:
 		return Float8, nil
 	case types.Decimal:
